server/ml-service/requirements.txt
flask
Pillow
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu